+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed using Composer "extras" instead of "extra".
+
 ## 1.0.1 - 2021-09-29
 
 ### Fixed